Fix Swift 6 compatibility issues
- Add import Accelerate to AudioChunker.swift
- Fix vDSP API calls (vDSP_Stride, FFTDirection, wrong parameter counts)
- Replace deprecated UnsafeMutablePointer.initialize(from:) with BufferPointer
- Add self. to property references in logger autoclosure context
- Fix Double/Float and Int/UInt32 type mismatches
- Replace deprecated preferredTrackRate with preferredRate
- Fix optional unwrapping (AudioMetadata, UnsafeMutablePointer)
- Change private members to internal for extension access
- Fix DispatchWorkItem trailing closure syntax
- Fix AudioQuality.lossless to AVAudioQuality.max
- Cast CMFormatDescription for streamBasicDescription access
- Add guard-let for optional AudioMetadata unwrapping

+++ ios/Classes/AudioChunker.swift
@@ -1,4 +1,5 @@
 import Foundation
+import Accelerate
 import AVFoundation
 import os.log
 
@@ -255,10 +256,10 @@ class AudioChunker: NSObject {
             totalChunks: chunks.count,
             totalDuration: totalDuration,
             processedDuration: processedDuration,
-            averageChunkDuration: chunks.isEmpty ? 0.0 : processedDuration / Double(chunks.count),
+            averageChunkDuration: chunks.isEmpty ? 0.0 : processedDuration / Float(chunks.count),
             speechChunks: chunks.filter { $0.hasSpeech }.count,
             silenceChunks: chunks.filter { $0.chunkType == .silence }.count,
-            processingProgress: totalDuration > 0 ? processedDuration / totalDuration : 0.0
+            processingProgress: totalDuration > 0 ? Float(processedDuration) / Float(totalDuration) : 0.0
         )
     }
